dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile dev-requirements.in
#
exceptiongroup==1.1.3
    # via pytest
faker==12.0.1
    # via mixer
flake8==6.1.0
    # via
    #   -r dev-requirements.in
    #   flake8-isort
    #   flake8-pyproject
    #   flake8-walrus
flake8-isort==6.1.0
    # via -r dev-requirements.in
flake8-pyproject==1.2.3
    # via -r dev-requirements.in
flake8-walrus==1.2.0
    # via -r dev-requirements.in
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via flake8-isort
mccabe==0.7.0
    # via flake8
mixer==7.2.2
    # via -r dev-requirements.in
packaging==23.2
    # via pytest
pluggy==1.3.0
    # via pytest
pycodestyle==2.11.0
    # via flake8
pyflakes==3.1.0
    # via flake8
pytest==7.4.2
    # via
    #   -r dev-requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r dev-requirements.in
python-dateutil==2.8.2
    # via faker
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via
    #   flake8-pyproject
    #   pytest
